requirements.txt
python-dotenv
discord.py
openai
tiktoken
wikipedia-api
